--- client/src/App.css
@@ -1,47 +1,0 @@
-.App {
-    text-align: center;
-  }
-  
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-    height: 80px;
-  }
-  
-  .App-header {
-    background-color: #222;
-    height: 150px;
-    padding: 20px;
-    color: white;
-  }
-  
-  .App-title {
-    font-size: 1.5em;
-  }
-  
-  .App-intro {
-    font-size: large;
-  }
-  
-  @keyframes App-logo-spin {
-    from { transform: rotate(0deg); }
-    to { transform: rotate(360deg); }
-  }
-  /* Change map position here */
-  #map {
-    height: 70vh;
-    width: 70vw;
-    margin: 20px 200px;
-  }
-  html, body {
-    height: 100%;
-    margin: 0;
-    padding: 0;
-  }
-@media screen and (max-width: 980px){
-  #map {
-    height: 260px;
-    width: 290px;
-    margin: 30px 20px;
-  }
-}
-
